% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{create_emission_matrix}
\alias{create_emission_matrix}
\title{Compute emission matrix}
\usage{
create_emission_matrix(ecm, alpha)
}
\arguments{
\item{ecm}{total emission count matrix}

\item{alpha}{smoothness coefficient}
}
\description{
Compute emission matrix
}
