% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{create_transition_matrix}
\alias{create_transition_matrix}
\title{Compute transition matrix}
\usage{
create_transition_matrix(tcm, alpha)
}
\arguments{
\item{tcm}{total transition count matrix on state space, which is tags space in POS}

\item{alpha}{smoothness coefficient}
}
\description{
Compute transition matrix
}
