% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Get dictionary}
\usage{
get_dictionary(words)
}
\arguments{
\item{words}{A list of words}
}
\value{
return a factor levels list for the given words, in POS they refer it to dictionary
}
\description{
Get dictionary
}
