% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penn_TreeBank.R
\docType{data}
\name{Penn_TreeBank}
\alias{Penn_TreeBank}
\title{Penn Treebank Sample Dataset}
\format{
A data frame with 100676 rows and 2 variables:
\describe{
  \item{word}{Character. The word token.}
  \item{tags}{Factor with levels corresponding to the Penn Treebank POS tags.
             Each level represents a particular part of speech.}
}
}
\source{
Original Penn Treebank Project
}
\usage{
Penn_TreeBank
}
\description{
A sample dataset extracted from the Penn Treebank, containing annotated
parts of speech for a selection of sentences. The dataset is widely used
for training and testing part-of-speech taggers and other natural language
processing tasks. Each entry consists of a word and its corresponding POS
tag as defined by the Penn Treebank project.
}
\references{
- Marcus, Mitchell P., et al. "Building a large annotated corpus of English:
  The Penn Treebank." Computational linguistics 19.2 (1993): 313-330.
- The Penn Treebank: https://www.cis.upenn.edu/~treebank/
}
\keyword{datasets}
