% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{plot_confusion_matrix_heatmap}
\alias{plot_confusion_matrix_heatmap}
\title{Plot Confusion Matrix as a Heatmap}
\usage{
plot_confusion_matrix_heatmap(
  conf_matrix,
  title = "Confusion Matrix",
  text_size = 5,
  plot_width = 12,
  plot_height = 12
)
}
\arguments{
\item{conf_matrix}{A confusion matrix object, typically created by caret::confusionMatrix() or similar.}

\item{title}{The title for the heatmap.}

\item{text_size}{A numerical value for the size of the text in the plot.}

\item{plot_width}{Width of the plot in inches, relevant for saving to a file.}

\item{plot_height}{Height of the plot in inches, relevant for saving to a file.}
}
\value{
A ggplot object representing the heatmap.
}
\description{
This function takes a confusion matrix and plots it as a heatmap. It adjusts the appearance of the heatmap for better readability, especially when the matrix is large. It allows for customization of text size and plot dimensions.
}
\examples{
# Example usage:
# Assuming 'cm' is a confusion matrix
# plot_confusion_matrix_heatmap(cm, title = "My Heatmap")
}
